+++ source/sites/publicodes/pages/Documentation.tsx
@@ -16,17 +16,15 @@ import {
 } from 'react-router-dom'
 import { RootState } from 'Reducers/rootReducer'
 import styled from 'styled-components'
-import { splitName } from '../../../components/publicodesUtils'
 import { useEngine } from '../../../components/utils/EngineContext'
-import Meta from '../../../components/utils/Meta'
 import { WithEngine } from '../../../RulesProvider'
 import { currentSimulationSelector } from '../../../selectors/storageSelectors'
-import { capitalise0 } from '../../../utils'
 import BandeauContribuer from '../BandeauContribuer'
 import RavijenChart from '../chart/RavijenChart'
 import References from '../DocumentationReferences'
 import DocumentationLanding from './DocumentationLanding'
 import DocumentationStyle from './DocumentationStyle'
+import QuickDocumentationPage from './QuickDocumentationPage'
 
 export default function () {
 	console.log('Rendering Documentation')
@@ -86,7 +84,7 @@ export default function () {
 
 			{!engineReady && !loadEngine && (
 				<div>
-					<QuickDocPage
+					<QuickDocumentationPage
 						rule={rule}
 						dottedName={dottedName}
 						setLoadEngine={setLoadEngine}
@@ -100,57 +98,6 @@ export default function () {
 			)}
 
 			<BandeauContribuer />
-		</div>
-	)
-}
-const QuickDocPage = ({ rule, dottedName, setLoadEngine }) => {
-	const split = splitName(dottedName),
-		title = rule.titre || capitalise0(split[splitName.length - 1]),
-		parents = split.slice(0, -1).join(' > ')
-
-	console.log(split, title)
-	return (
-		<div
-			css={`
-				max-width: calc(800px + 1.2rem);
-				margin: 0 auto;
-			`}
-		>
-			<DocumentationStyle>
-				<Meta description={rule.description} title={title} />
-				<header>
-					<small>{parents}</small>
-					<h1>{title}</h1>
-				</header>
-				<section>
-					{rule.description && <Markdown>{rule.description}</Markdown>}
-				</section>
-				<button
-					onClick={() => setLoadEngine(true)}
-					className="ui__ button cta plain"
-				>
-					🧮 Lancer le calcul
-				</button>
-				{rule.formule && (
-					<div>
-						<h2>Comment cette donnée est-elle calculée ?</h2>
-
-						<blockquote>{JSON.stringify(rule.formule, null, 3)}</blockquote>
-					</div>
-				)}
-				{rule.note && (
-					<div>
-						<h2>Notes</h2>
-						<Markdown>{rule.note}</Markdown>
-					</div>
-				)}
-				{rule.références && (
-					<div>
-						<h2>Références</h2>
-						<References references={rule.références} />
-					</div>
-				)}
-			</DocumentationStyle>
 		</div>
 	)
 }
